Implement `d2d` and `y2y`
This ended up being a pretty large commit as implementing these two
sequences uncovered a weakness in the NumberPrefix operator. Not all
operations can be simply repeated. In yank's case it would just copy a
single line instead of two.

Therefore I changed the select method, to always take a count and return an
array of whether the selection was successful. While this requires more
typing I believe it better communicates to future committers that they
need to consider how their new operation or motion works in a repetitive
manner.