Comment out tests for `-` that are failing due to `k`
I don’t want to delete these tests because they will be useful to have when `k` (and correspondingly `j`) is fixed. Otherwise, they would have to be rewritten at that time, or they would not get written and `-` and `+` would not be as well-tested.

The comments explaining why these tests are commented out and saying when they can be re-enabled should protect these tests from being left disabled when they shouldn’t be.