Add tests for + and enter, mirroring the tests for -
As requested [in the pull request discussion](https://github.com/atom/vim-mode/pull/277#issuecomment-47765173), I am adding the tests for + even though some of the copied tests are disabled.

These tests are all the same as the ones for -, except that the starting line numbers are X from the bottom instead of X from the top, and in the “abc” sample text the order of the lines is reversed.

I also changed a previous test for - so that its cursor column is less remarkable, and won’t fail due to edge cases that that particular test is not aiming to test. The cursor column in that test really is now in the middle of the line, instead of on the last character.