Add tests for enter, checking equivalence to +
I added tests for enter too, for the heck of it. To avoid duplication, the tests merely compare the results of + and of enter and make sure that they are the same.

I could have tested equivalence of enter and + in more situations, but I think it’s likely that if they are the same in the situation tested, they will be the same anywhere. It would be an unlikely mistake for someone to change the keybindings of enter and + to different functions, and then edit only one of the functions in such a way that the results are only sometimes different.

I also changed some single quotes to double quotes for consistency.