Comment out only the failing assertions in the - and + tests
The behavior of `j` and `k` have changed since the tests were written. The commented-out tests still fail, but not as badly. The cursor column is still wrong, but the cursor row and the text in the editor are now correct.

Thus, so that the text in the editor can be automatically tested, the tests have been uncommented, and only the screen position check has been commented out.