Fix scrolling up/down so it moves by 1 line.
Unfortunately, atom is not showing the integer number of lines all the time
so sometimes it looks like it's not scrolling correctly. This is because
scrolling up (ctrl-y) is done by requesting atom to show the top lines and
so very bottom line may be only partially shown (and so it sometimes feels
like it does not work correctly).