:fix: repeating insert operations
An `I` or `A` insert operation should be repeated as such, rather than
as a simple `i` or `a` operation.
Fixing this led me to find a bug: repeated insert didn’t position the
cursor correctly at the last inserted character (not after it), so even
repeated `a` or `i` wouldn’t work right; this is fixed too.