Adjust replace mode backspace to work w/ new history API
As of atom/text-buffer#81, which will ship in Atom v1.0.3, the editor's history
API has changed in some subtle ways in order to preserve the proper nesting of
transactions. The `groupChangesSinceCheckpoint` method no longer works if called
from within a transaction with a checkpoint that was created *outside* the
transaction.

/cc @jacekkopecky, can you double check that this alternative approach works
correctly?