merge cursors in ensure...
In the unusual case that a line already has a cursor on the last
character and another comes to the end of the line and
`ensureCursorsWithinLine` moves it, we can end up with two cursors in
the same place. That’s undesirable.

This can happen for example when moving only one of multiple cursors
(see joseramonc/multi-cursor#4).