:bug: Fix repeat-insertion bug referenced in #1006
As reported several times in issue #1006, the dot-command
no longer repeats insertions since roughly Atom release 1.7.0.
This can be traced back to an update in the text-buffer package,
which appears to have changed the behavior of
getChangesSinceCheckpoint in lib/vim-state.coffee and the
TransactionBuilder class in input.coffee which uses its return.

This commit replaces the local getChangesSinceCheckpoint function
by the counterpart of the same name from the text-buffer package,
rendering the local implementation obsolete. The TransactionBuilder
class is rewritten to deal with this alternate input.